[Store] Revert group test split; keep group behavior tests in master_service_test
