+++ mooncake-store/tests/master_service_test.cpp
@@ -6534,6 +6534,257 @@ TEST_F(MasterServiceTest, GracefulUnmountSegment_PreventAllocation) {
     std::this_thread::sleep_for(std::chrono::milliseconds(1500));
 }
 
+TEST_F(MasterServiceTest, GroupedRoutingUsesHashOfTenantAndKeyOnly) {
+    // Route-decoupling invariant: object routing is a pure function of
+    // (tenant, key); the group_id is only a lifecycle annotation and never
+    // affects which metadata shard an object lands in. The group domain is
+    // keyed by scoped(tenant, group_id) and stores only the member key list.
+    std::unique_ptr<MasterService> service_(new MasterService());
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    // Two member keys that hash to different metadata shards, sharing one
+    // group whose id hashes to yet another shard. The default-tenant route is
+    // hash(key) % kNumShards (mirrors MasterService::getShardIndex).
+    constexpr size_t kMetadataShardCountForTest = 1024;
+    const std::string key_a = "route_decouple_key_a";
+    const std::string group_id = FindGroupIdOnDifferentShard(key_a);
+    std::string key_b = "route_decouple_key_b";
+    const size_t shard_a =
+        std::hash<std::string>{}(key_a) % kMetadataShardCountForTest;
+    size_t shard_b =
+        std::hash<std::string>{}(key_b) % kMetadataShardCountForTest;
+    for (int i = 0; i < 10000 && shard_b == shard_a; ++i) {
+        key_b = "route_decouple_key_b_" + std::to_string(i);
+        shard_b = std::hash<std::string>{}(key_b) % kMetadataShardCountForTest;
+    }
+    ASSERT_NE(shard_a, shard_b);  // members span metadata shards
+
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.group_ids = std::vector<std::string>{group_id};
+    PutCompletedObject(*service_, client_id, key_a, config);
+    PutCompletedObject(*service_, client_id, key_b, config);
+
+    // Both members are reachable purely through hash(tenant, key) routing,
+    // which is decoupled from the group domain.
+    EXPECT_TRUE(service_->ExistKey(key_a, TenantId::Default()).value_or(false));
+    EXPECT_TRUE(service_->ExistKey(key_b, TenantId::Default()).value_or(false));
+    EXPECT_TRUE(
+        service_->GetReplicaList(key_a, TenantId::Default()).has_value());
+    EXPECT_TRUE(
+        service_->GetReplicaList(key_b, TenantId::Default()).has_value());
+
+    // The group table still sees both members: group state is a separate,
+    // key-list-only domain.
+    auto members = GetGroupMemberKeysForTest(*service_, group_id);
+    EXPECT_EQ(2u, members.size());
+
+    // Route stability: the object route is hash(tenant, key) alone — grouping
+    // does not change it (identical to the ungrouped route computed before the
+    // objects existed), so a later ungrouped put of the same key would land on
+    // the same shard.
+    EXPECT_EQ(shard_a,
+              std::hash<std::string>{}(key_a) % kMetadataShardCountForTest);
+    EXPECT_EQ(shard_b,
+              std::hash<std::string>{}(key_b) % kMetadataShardCountForTest);
+}
+
+TEST_F(MasterServiceTest, GroupedReadRefreshesSharedGroupLease) {
+    auto service_config =
+        MasterServiceConfig::builder().set_default_kv_lease_ttl(200).build();
+    std::unique_ptr<MasterService> service_(new MasterService(service_config));
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    const std::string key_a = "lease_group_key_a";
+    const std::string key_b = "lease_group_key_b";
+    const std::string group_id = FindGroupIdOnDifferentShard(key_a);
+
+    ReplicateConfig config_a;
+    config_a.replica_num = 1;
+    config_a.group_ids = std::vector<std::string>{group_id};
+    ReplicateConfig config_b = config_a;
+
+    PutCompletedObject(*service_, client_id, key_a, config_a);
+    PutCompletedObject(*service_, client_id, key_b, config_b);
+
+    // Read key_a (twice, near expiry). Group protection is keyed on ONE shared
+    // group TTL, so reading a member refreshes the group TTL and protects the
+    // WHOLE group (both key_a and key_b), not just the read member.
+    auto exists = service_->ExistKey(key_a, TenantId::Default());
+    ASSERT_TRUE(exists.has_value());
+    ASSERT_TRUE(exists.value());
+
+    std::this_thread::sleep_for(std::chrono::milliseconds(120));
+    exists = service_->ExistKey(key_a, TenantId::Default());
+    ASSERT_TRUE(exists.has_value());
+    ASSERT_TRUE(exists.value());
+
+    // The shared group TTL is active, so a non-force remove of key_a is
+    // rejected.
+    auto remove_read_member = service_->Remove(key_a, TenantId::Default());
+    ASSERT_FALSE(remove_read_member.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_read_member.error());
+
+    // key_b shares the same group TTL and is therefore ALSO protected.
+    auto remove_peer = service_->Remove(key_b, TenantId::Default());
+    ASSERT_FALSE(remove_peer.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_peer.error());
+
+    // Force cleanup both members.
+    EXPECT_TRUE(service_->Remove(key_a, TenantId::Default(), /*force=*/true)
+                    .has_value());
+    EXPECT_TRUE(service_->Remove(key_b, TenantId::Default(), /*force=*/true)
+                    .has_value());
+}
+
+TEST_F(MasterServiceTest, GroupedMembershipChangeStillSharesGroupLeaseOnRead) {
+    auto service_config =
+        MasterServiceConfig::builder().set_default_kv_lease_ttl(500).build();
+    std::unique_ptr<MasterService> service_(new MasterService(service_config));
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    const std::string key_a = "lease_group_dirty_key_a";
+    const std::string key_b = "lease_group_dirty_key_b";
+    const std::string group_id = FindGroupIdOnDifferentShard(key_a);
+
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.group_ids = std::vector<std::string>{group_id};
+
+    PutCompletedObject(*service_, client_id, key_a, config);
+    ASSERT_TRUE(service_->ExistKey(key_a, TenantId::Default()).value_or(false));
+
+    // Add a new member to the group after key_a was already written. On put,
+    // key_b is wired to the SAME shared Lease, so it participates in the
+    // group's protection.
+    PutCompletedObject(*service_, client_id, key_b, config);
+    std::this_thread::sleep_for(std::chrono::milliseconds(150));
+
+    // Reading key_a refreshes the shared group TTL, which key_b also shares.
+    auto exists = service_->ExistKey(key_a, TenantId::Default());
+    ASSERT_TRUE(exists.has_value());
+    ASSERT_TRUE(exists.value());
+
+    // key_a is protected by the shared group lease...
+    auto remove_read_member = service_->Remove(key_a, TenantId::Default());
+    ASSERT_FALSE(remove_read_member.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_read_member.error());
+
+    // ...and key_b (sharing the same group TTL) is protected too.
+    auto remove_peer = service_->Remove(key_b, TenantId::Default());
+    ASSERT_FALSE(remove_peer.has_value());
+    EXPECT_EQ(ErrorCode::OBJECT_HAS_LEASE, remove_peer.error());
+
+    EXPECT_TRUE(service_->Remove(key_a, TenantId::Default(), /*force=*/true)
+                    .has_value());
+    EXPECT_TRUE(service_->Remove(key_b, TenantId::Default(), /*force=*/true)
+                    .has_value());
+}
+
+TEST_F(MasterServiceTest, GroupStateRegistersAndCleansUpMembers) {
+    std::unique_ptr<MasterService> service_(new MasterService());
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    const std::string key_a = "group_state_key_a";
+    const std::string key_b = "group_state_key_b";
+    const std::string group_id = "group_state_group";
+
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.group_ids = std::vector<std::string>{group_id};
+    PutCompletedObject(*service_, client_id, key_a, config);
+    PutCompletedObject(*service_, client_id, key_b, config);
+
+    auto members = GetGroupMemberKeysForTest(*service_, group_id);
+    EXPECT_EQ(2u, members.size());
+
+    // Removing one member shrinks, but does not erase, the group.
+    ASSERT_TRUE(service_->Remove(key_a, TenantId::Default(), /*force=*/true)
+                    .has_value());
+    members = GetGroupMemberKeysForTest(*service_, group_id);
+    ASSERT_EQ(1u, members.size());
+    EXPECT_EQ(key_b, members[0]);
+
+    // Removing the last member erases the group.
+    ASSERT_TRUE(service_->Remove(key_b, TenantId::Default(), /*force=*/true)
+                    .has_value());
+    EXPECT_TRUE(GetGroupMemberKeysForTest(*service_, group_id).empty());
+}
+
+TEST_F(MasterServiceTest, RebuildGroupStateRestoresMembershipFromMetadata) {
+    std::unique_ptr<MasterService> service_(new MasterService());
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    const std::string key_a = "rebuild_group_key_a";
+    const std::string key_b = "rebuild_group_key_b";
+    const std::string group_id = "rebuild_group_id";
+
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.group_ids = std::vector<std::string>{group_id};
+    PutCompletedObject(*service_, client_id, key_a, config);
+    PutCompletedObject(*service_, client_id, key_b, config);
+
+    // Simulate a snapshot reset: drop all group state.
+    ClearGroupStateForTest(*service_);
+    EXPECT_TRUE(GetGroupMemberKeysForTest(*service_, group_id).empty());
+
+    // Rebuild from object metadata (as snapshot deserialization does).
+    RebuildGroupStateForTest(*service_);
+
+    auto members = GetGroupMemberKeysForTest(*service_, group_id);
+    EXPECT_EQ(2u, members.size());
+}
+
+TEST_F(MasterServiceTest, GroupLeaseIsSharedAndExtendsOnMemberRead) {
+    const uint64_t kv_lease_ttl = 1000;
+    auto service_config = MasterServiceConfig::builder()
+                              .set_default_kv_lease_ttl(kv_lease_ttl)
+                              .build();
+    std::unique_ptr<MasterService> service_(new MasterService(service_config));
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    const UUID client_id = generate_uuid();
+
+    const std::string key_a = "group_lease_member_a";
+    const std::string key_b = "group_lease_member_b";
+    const std::string group_id = FindGroupIdOnDifferentShard(key_a);
+
+    ReplicateConfig config;
+    config.replica_num = 1;
+    config.group_ids = std::vector<std::string>{group_id};
+    PutCompletedObject(*service_, client_id, key_a, config);
+    PutCompletedObject(*service_, client_id, key_b, config);
+
+    // Both members resolve to the SAME shared Lease.
+    const auto lease_a = GetGroupLeaseForTest(*service_, group_id);
+    ASSERT_NE(nullptr, lease_a);
+    const auto lease_b = GetGroupLeaseForTest(*service_, group_id);
+    ASSERT_NE(nullptr, lease_b);
+    EXPECT_EQ(lease_a.get(), lease_b.get());
+
+    // A freshly-created group (no reads yet) is not protected -> evictable.
+    EXPECT_TRUE(lease_a->IsExpired(std::chrono::system_clock::now()));
+
+    // Reading one member extends the shared group lease -> the whole group is
+    // now protected.
+    EXPECT_TRUE(
+        service_->GetReplicaList(key_a, TenantId::Default()).has_value());
+    EXPECT_FALSE(lease_a->IsExpired(std::chrono::system_clock::now()));
+    // The other member sees the same (shared) extended deadline.
+    EXPECT_FALSE(lease_b->IsExpired(std::chrono::system_clock::now()));
+}
+
+TEST_F(MasterServiceTest, ReRouteRestoredObjectsMovesStaleShardObjects) {
+    std::unique_ptr<MasterService> service_(new MasterService());
+    [[maybe_unused]] const auto context = PrepareSimpleSegment(*service_);
+    ReRouteRestoredObjectsMigrationForTest(*service_);
+}
+
 TEST_F(MasterServiceTest, RemoveAllLeasedObject) {
     const uint64_t kv_lease_ttl = 50;
     auto service_config = MasterServiceConfig::builder()
